Rename device.py to hub.py
