+++ homeassistant/components/axis/hub.py
